Replace f32 by f64

+++ src/hlr_lowering.rs
@@ -168,7 +168,7 @@ impl<'a, 'ctxt: 'a> HlrLowerer<'a, 'ctxt> {
                 (mlr::Const::Int(*n), ty)
             }
             hlr::Lit::Float(f) => {
-                let ty = self.builder.ctxt.tys.primitive(ty::Primitive::Float32);
+                let ty = self.builder.ctxt.tys.primitive(ty::Primitive::Float64);
                 (mlr::Const::Float(*f), ty)
             }
             hlr::Lit::Bool(b) => {
