+++ src/parse.rs
@@ -802,7 +802,7 @@ impl<'ast, 'token> AstParser<'ast, 'token> {
             }
             Token::FloatLiteral(int_part, frac_part) => {
                 let combined = format!("{int_part}.{frac_part}");
-                let value: f32 = combined.parse().map_err(|_| ParserErr::InvalidLiteral)?;
+                let value: f64 = combined.parse().map_err(|_| ParserErr::InvalidLiteral)?;
                 self.tokens.advance();
                 Ok(self.builder.lit(Lit::Float(value)))
             }
